link table as a table actions in the inspector
