+++ mathesar_ui/src/systems/table-view/table-inspector/TableActions.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { Button, Icon } from '@mathesar/component-library';
-  import { iconDelete, iconQuery } from '@mathesar/icons';
+  import { iconDelete, iconQuery, iconTableLink } from '@mathesar/icons';
   import { confirmDelete } from '@mathesar/stores/confirmation';
   import { getTabularDataStoreFromContext } from '@mathesar/stores/table-data';
   import { deleteTable, refetchTablesForSchema } from '@mathesar/stores/tables';
@@ -8,9 +8,13 @@
   import { currentSchemaId } from '@mathesar/stores/schemas';
   import { getDataExplorerPageUrl } from '@mathesar/routes/urls';
   import { currentDatabase } from '@mathesar/stores/databases';
+  import { modal } from '@mathesar/stores/modal';
+  import LinkTableModal from '../link-table/LinkTableModal.svelte';
 
   const dispatch = createEventDispatcher();
   const tabularData = getTabularDataStoreFromContext();
+  const linkTableModal = modal.spawnModalController();
+  const tableConstraintsModal = modal.spawnModalController();
 
   function handleDeleteTable() {
     void confirmDelete({
@@ -28,6 +32,15 @@
 </script>
 
 <div class="actions-container">
+  <Button appearance="ghost" on:click={() => linkTableModal.open()}>
+    <Icon {...iconTableLink} />
+    <span>Link Table</span>
+  </Button>
+  <LinkTableModal
+    controller={linkTableModal}
+    on:goToConstraints={() => tableConstraintsModal.open()}
+  />
+
   {#if $currentDatabase && $currentSchemaId}
     <Button appearance="ghost">
       <Icon {...iconQuery} />
